application/requirements.txt
uvicorn
fastapi
fastapi-utils
jinja2
python-multipart
colorama
pytz
requests
python-dotenv